requirements.txt
streamlit
langchain>=0.1.14
langchain-community>=0.0.25
langchain-openai>=0.1.1
portkey-ai>=0.1.18
tavily-python>=0.3.0
chromadb>=0.4.22
PyYAML>=6.0
pypdf>=3.16.4